+++ packages/ui/src/utils/propertyMap.js
@@ -16,7 +16,9 @@ export function convertNumberStyleValue(key, value) {
 export function propertyMap(styles = {}, mapper = {}) {
   const styleObj = {};
   Object.keys(styles).forEach((key) => {
-    styleObj[mapper[key] || key] = convertNumberStyleValue(key, styles[key]);
+    if (typeof styles[key] !== "undefined") {
+      styleObj[mapper[key] || key] = convertNumberStyleValue(key, styles[key]);
+    }
   });
   return styleObj;
 }